refactor: split the code into the correct structure and adding components to a separate folder
