+++ rs-react-app/src/App.module.css
@@ -1,6 +1,34 @@
 .wrapper {
 }
 
+.topControls {
+  display: flex;
+  gap: 10px;
+  max-width: 900px;
+  margin: 0 auto;
+  padding-top: 20px;
+  input,
+  button {
+    height: 35px;
+    border: 1px solid rgb(156, 156, 156);
+    border-radius: 4px;
+  }
+}
+
+.topControls button {
+  padding: 5px 15px;
+  font-family: 'Orbitron', sans-serif;
+  text-transform: uppercase;
+  font-weight: 600;
+  letter-spacing: 0.01em;
+}
+
+.topControls input {
+  padding: 5px;
+  width: 50%;
+  font-size: 20px;
+}
+
 .cardsContainer {
   max-width: 900px;
   margin: 0 auto;
@@ -16,19 +44,4 @@
   justify-content: center;
   align-items: center;
   z-index: 100;
-}
-
-.spinner {
-  width: 50px;
-  height: 50px;
-  border: 5px solid rgba(0, 0, 0, 0.1);
-  border-radius: 50%;
-  border-top-color: #0092f4;
-  animation: spin 1s ease infinite;
-}
-
-@keyframes spin {
-  to {
-    transform: rotate(360deg);
-  }
 }
