+++ rs-react-app/src/components/card/Card.module.css
@@ -1,5 +1,8 @@
+@import url('https://fonts.googleapis.com/css2?family=Orbitron:wght@400..900&display=swap');
+
 .card {
   display: flex;
+  font-family: 'Orbitron', sans-serif;
   align-items: center;
   gap: 60px;
   padding: 10px;
