+++ rs-react-app/src/components/spinner/Spinner.tsx
@@ -1,0 +1,10 @@
+import { Component } from 'react';
+import styles from './spinner.module.css';
+
+class Spinner extends Component {
+  render() {
+    return <div className={styles.spinner}></div>;
+  }
+}
+
+export default Spinner;
