+++ rs-react-app/src/components/spinner/index.ts
@@ -1,0 +1,3 @@
+import Spinner from './Spinner';
+
+export default Spinner;
