+++ rs-react-app/src/utils/formatKey.ts
@@ -1,0 +1,5 @@
+export default function formatKey(key: string): string {
+  return key
+    .replace(/([A-Z])/g, ' $1')
+    .replace(/^./, (str) => str.toUpperCase());
+}
