+++ rs-react-app/vite.config.ts
@@ -1,12 +1,17 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 
-// https://vite.dev/config/
 export default defineConfig({
   css: {
     modules: {
       localsConvention: 'camelCase',
     },
+  },
+  base: './',
+  build: {
+    outDir: 'dist',
+    assetsDir: 'assets',
+    emptyOutDir: true,
   },
   plugins: [react()],
 });
